Start chromedriver once for running all integration tests
